Introduce Ingredients and Ingredient models
Allows for the use of separate ingredient lists for parts of a recipe
(like the sauce on a pizza or whathaveyou). Ingredient is split up in
qunatity, unit and name to allow for future automatic multiplication
of quantities for a requested number of people.

The current admin is hard to use as it involves the creation of many
intermediate models to make up the ingredient listings for a single
recipe. That should change in the future :P

Heavily inspired by @ekeijl's approach in PR #4.

This commit also goes towards #3.